[SITIONIX-126] - extract chat lifecycle mapping to dedicated mapstruct mappers
