+++ api-rest/src/main/java/com/sitionix/bffssox/mapper/ChatAgentApiMapper.java
@@ -6,24 +6,25 @@ import com.app_afesox.bffssox.api_first.dto.AgentConversationMessageDTO;
 import com.app_afesox.bffssox.api_first.dto.AgentConversationsResponseDTO;
 import com.app_afesox.bffssox.api_first.dto.ChatAgentRequestDTO;
 import com.app_afesox.bffssox.api_first.dto.ChatExecutionDTO;
-import com.app_afesox.bffssox.api_first.dto.ChatExecutionFailureDTO;
 import com.app_afesox.bffssox.api_first.dto.ChatAgentResponseDTO;
-import com.app_afesox.bffssox.api_first.dto.ExecutionStatusDTO;
 import com.app_afesox.bffssox.api_first.dto.SubmitChatExecutionResponseDTO;
 import com.sitionix.bffssox.domain.AgentConversation;
 import com.sitionix.bffssox.domain.AgentConversationDetails;
 import com.sitionix.bffssox.domain.AgentConversationsResponse;
 import com.sitionix.bffssox.domain.ChatExecution;
-import com.sitionix.bffssox.domain.ChatExecutionFailure;
 import com.sitionix.bffssox.domain.ChatAgentMessage;
 import com.sitionix.bffssox.domain.ChatAgentRequest;
 import com.sitionix.bffssox.domain.ChatAgentResponse;
 import com.sitionix.bffssox.domain.SubmitChatExecutionResponse;
 import java.util.List;
+import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mapper;
 
-@Mapper(componentModel = "spring")
+@Mapper(componentModel = "spring", injectionStrategy = InjectionStrategy.CONSTRUCTOR, uses = {
+        ChatExecutionStatusApiMapper.class,
+        ChatExecutionFailureApiMapper.class
+})
 public interface ChatAgentApiMapper {
 
     ChatAgentRequest asChatAgentRequest(ChatAgentRequestDTO src);
@@ -67,25 +68,4 @@ public interface ChatAgentApiMapper {
         return value == null ? null : AgentConversationMessageDTO.AuthorTypeEnum.fromValue(value);
     }
 
-    default ExecutionStatusDTO mapExecutionStatus(final String value) {
-        if (value == null) {
-            return null;
-        }
-        return switch (value) {
-            case "QUEUED" -> ExecutionStatusDTO.ACCEPTED;
-            case "COMPLETED" -> ExecutionStatusDTO.SUCCEEDED;
-            default -> ExecutionStatusDTO.fromValue(value);
-        };
-    }
-
-    default ChatExecutionFailureDTO asChatExecutionFailureDto(final ChatExecutionFailure src) {
-        if (src == null) {
-            return null;
-        }
-        return ChatExecutionFailureDTO.builder()
-                .code(src.getFailureClass())
-                .message(src.getReason())
-                .details(src.getRetryable() == null ? null : java.util.Map.of("retryable", src.getRetryable()))
-                .build();
-    }
 }
