+++ clients/client-atmssox/src/main/java/com/sitionix/bffssox/mapper/ChatExecutionFailureClientMapper.java
@@ -1,0 +1,21 @@
+package com.sitionix.bffssox.mapper;
+
+import com.app_afesox.atmssox.api_first.dto.ChatExecutionFailureDTO;
+import com.sitionix.bffssox.domain.ChatExecutionFailure;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ChatExecutionFailureClientMapper {
+
+    default ChatExecutionFailure asChatExecutionFailure(final ChatExecutionFailureDTO src) {
+        if (src == null) {
+            return null;
+        }
+        final Object retryable = src.getDetails() == null ? null : src.getDetails().get("retryable");
+        return ChatExecutionFailure.builder()
+                .failureClass(src.getCode())
+                .reason(src.getMessage())
+                .retryable(retryable instanceof Boolean value ? value : null)
+                .build();
+    }
+}
